@@ -2,27 +2,31 @@
   "name": "esbuild-ts-starter",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
+  "module": "node18",
   "scripts": {
     "start": "node ./dist",
-    "dev": "concurrently \"pnpm build:watch\" \"nodemon dist/index.js\"",
     "build": "node scripts/build.js",
-    "build:watch": "node scripts/build.js -watch",
+    "build:watch": "node scripts/build.js -w",
+    "dev": "node scripts/build.js -d",
     "test": "jest",
     "test:watch": "jest --watch",
-    "clean": "rimraf dist coverage"
+    "clean": "rimraf dist coverage",
+    "script:build": "esbuild --outdir=scripts --target=node18 scripts/build.ts"
   },
   "dependencies": {
-    "@types/jest": "^29.2.3",
     "lodash": "^4.17.21",
     "moment": "^2.29.4"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.1",
     "@types/lodash": "^4.14.182",
-    "concurrently": "^7.6.0",
+    "@types/node": "18.15.3",
+    "@types/nodemon": "^1.19.2",
     "esbuild": "^0.17.19",
     "jest": "^29.3.1",
     "nodemon": "^2.0.20",
     "ts-jest": "^29.0.3",
-    "typescript": "4.9.3"
+    "typescript": "^5.0.4"
   }
 }
